parse tokens - add support for i32->number
The Integer32 in our type will be captured with the ident value.
we are only focusing on our simple type for the time being and ignoring all other matches
If any type that our utility does not know how to process is encountered, the dbg! macro will print a message along with a convenient line number so user's can easily identify which condition was hit and expand on the utility if they choose.

All of Rust's many different types of numbers will simply be treated as a number when deserialized as Typescript`

refactor - add support for integers 32